split out compiler
